Added ChessRefactor
Removed useless files, decoupled classes and functionalities.
Used proper structure for project in a way that it would be easy
to scale it, update, maintain and so on.

While making for example two Chess games simultaneously would be
impossible on initial release, refactor focused on making changes
and navigating more accessible.

Code is mostly self documented and organized.